Accepting tech debt and adding built KAR.
The ADT connector is highly configurable and until now we have never had a situation that has required us to modify the source code. Therefore we are taking the pragmatic decision to simply upload the built KAR that we have been using so that it can be easily integrated into the CI pipeline (using Docker).

If in future we need to make some modifications we understand that we first need to resolve the tech debt so that the KAR is rebuilt on every CI run.