Made Redis host dynamic so can be run on GoCD.
Our CI pipline uses GoCD to run tests on GoCD Agents. When the agents are asked to run Docker containers it does not run them locally (not sure how or why, ask our Platform Engineer).
A Docker compose file is used to bring up a couple of containers that the Maven test depend on. The Maven tests are then run outside of Docker.
When running locally this works fine as the test look at localhost. It does not work on the Go Agents where the docker containers are somewhere else.

The solution is quite horrible. It leverages the properties mechanism already being used to substitute values into the beans XML configuration.
I've added a property for the Redis host which is itself a placeholder. The placeholder is replaced using `sed` in the Makefile.