PR builds now publishes images.
PR builds themselves don't really have a need to publish images but it is for the benefit of builds triggered from merges into master.
In order to not duplicate all the effort of building and testing, when a merge into master occurs, the repository can be checked for the already verified image and simply be given extra tags.
Pushes